cb namespace removal / build fix
